Change descriptions in Tooltip.xml
